Only deletes an iteration if there are already 10

+++ Trunked/Trunked/CustomVision.cs
@@ -119,7 +119,8 @@ namespace Trunked
 
             Iteration iterationToDelete = iterations[iterations.Count - 1];
 
-            TrainingApi.DeleteIteration(ProjectID, iterationToDelete.Id);
+            if (iterations.Count == 10)
+                TrainingApi.DeleteIteration(ProjectID, iterationToDelete.Id);
         }
     }
 }
